More waypoint editing

+++ src/Web/App.ts
@@ -167,7 +167,7 @@ export class App {
 
   handleEventClickWaypointEdit(target: HTMLButtonElement) {
     const type = target.getAttribute("data-type");
-    const waypointId = Number((target.closest("dialog") as HTMLDialogElement).getAttribute("data-waypoint-id")) - 1;
+    const waypointId = Number((target.closest("dialog") as HTMLDialogElement).getAttribute("data-cp-id"));
     switch (type) {
       case "delete":
         this.mission.checkpoints.splice(waypointId, 1);
@@ -406,7 +406,7 @@ export class App {
         break;
       default:
         const prop = target.getAttribute("data-cp-prop");
-        const id = target.getAttribute("data-cp-id");
+        const id = (target.closest("[data-cp-id]") || target).getAttribute("data-cp-id");
         if (prop && id) {
           const index = Number(id);
           switch (prop) {
@@ -425,6 +425,14 @@ export class App {
               break;
             case "altitude_ft":
               this.mission.checkpoints[index].lon_lat.altitude_ft = target.valueAsNumber;
+              break;
+            case "lat":
+              this.mission.checkpoints[index].lon_lat.lat = target.valueAsNumber;
+              show |= App.SHOW_MAP;
+              break;
+            case "lon":
+              this.mission.checkpoints[index].lon_lat.lon = target.valueAsNumber;
+              show |= App.SHOW_MAP;
               break;
             case "speed":
               this.mission.checkpoints[index].speed = target.valueAsNumber;
@@ -619,13 +627,27 @@ export class App {
             return;
           }
           const modal = document.getElementById("edit-waypoint-modal") as HTMLDialogElement;
-          modal.setAttribute("data-waypoint-id", String(currentFeature.id));
+          const currentCheckpointIndex = Number(currentFeature.id) - 1;
+          if (currentCheckpointIndex < 0 || currentCheckpointIndex >= this.mission.checkpoints.length) {
+            return;
+          }
+          const currentCheckpoint = this.mission.checkpoints[currentCheckpointIndex];
+          modal.setAttribute("data-cp-id", String(currentCheckpointIndex));
           (modal.querySelector('[data-type="delete"]') as HTMLButtonElement).disabled =
             currentFeature.id === 1 || currentFeature.id === this.mission.checkpoints.length - 1;
           (modal.querySelector('[data-type="add-before"]') as HTMLButtonElement).disabled = currentFeature.id === 1;
           (modal.querySelector('[data-type="add-after"]') as HTMLButtonElement).disabled =
             currentFeature.id === this.mission.checkpoints.length;
-          console.log(currentFeature.id, modal.querySelector('[data-type="add-after"]') as HTMLButtonElement);
+          (document.getElementById("wp-lon") as HTMLInputElement).value = currentCheckpoint.lon_lat.lon.toFixed(5);
+          (document.getElementById("wp-lat") as HTMLInputElement).value = currentCheckpoint.lon_lat.lat.toFixed(5);
+          (document.getElementById("wp-name") as HTMLInputElement).value = currentCheckpoint.name;
+          modal.addEventListener(
+            "close",
+            (e) => {
+              this.showFlightplan(App.SHOW_ALL);
+            },
+            { once: true }
+          );
           modal.showModal();
           return;
         }
@@ -742,15 +764,19 @@ export class App {
     });
     modal.showModal();
 
-    (modal.querySelector("button") as HTMLButtonElement).addEventListener("click", (e) => {
-      e.stopPropagation();
-      e.preventDefault();
-      new MissionFactory().create(mlp.getMissionString(Number(select.value)), this.mission);
-      this.useIcao = this.mission.origin_country !== "US";
-      this.syncToForm();
-      this.showFlightplan(App.SHOW_ALL | App.SHOW_MAP_CENTER);
-      modal.close();
-    });
+    (modal.querySelector("button") as HTMLButtonElement).addEventListener(
+      "click",
+      (e) => {
+        e.stopPropagation();
+        e.preventDefault();
+        new MissionFactory().create(mlp.getMissionString(Number(select.value)), this.mission);
+        this.useIcao = this.mission.origin_country !== "US";
+        this.syncToForm();
+        this.showFlightplan(App.SHOW_ALL | App.SHOW_MAP_CENTER);
+        modal.close();
+      },
+      { once: true }
+    );
   }
 
   makeWeather() {
